Autotag using a PAT instead of github.token
Using github.token means that the action will not trigger any further
workflows. However, what we want is for workflows to be triggered. So we
use a PAT here to ensure that happens.

This is secure since PRs only have access to secrets if the PR is made
from a branch within the same repository.